Tables signatures and relationships added.

+++ database/migrations/2021_04_10_212513_create_websites_table.php
@@ -16,7 +16,7 @@ class CreateWebsitesTable extends Migration
         Schema::create('websites', function (Blueprint $table) {
             $table->increments('id');
 
-            $table->bigInteger('author_id')
+            $table->integer('author_id')
                 ->constrained()
                 ->onDelete('restrict');
